pyproject.toml
[tool.poetry]
name = "harmony"
version = "0.1.0"
description = ""
authors = ["Tommaso Bianco <tommaso.bianco@gmail.com>"]
readme = "README.md"
packages = [{include = "harmony", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.28.1"
dacite = "^1.8.1"
python-dotenv = "^1.0.0"
tiktoken = "^0.5.1"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
